# DRY sampler performance optimization