## Background

The DRY sampler by @p-e-w was [recently merged](https://github.com/oobabooga/text-generation-webui/pull/5677) to dev. It seems to be a significant improvement in the fight against repetition, but there were concerns about the performance impact of the sampler. When @oobabooga merged the PR, they [requested ideas to make it faster](https://github.com/oobabooga/text-generation-webui/pull/5677#issuecomment-2119572532), so I decided to tackle this as my first contribution to text-generation-webui.

## Performance impact of this PR

- The "worst case" scenario for the original DRY implementation is an input which contains nothing else except the same token repeated many times. **A repeating input of 1024 tokens takes 38 seconds to run on my machine with the original implementation, compared to ~0.00 seconds with my optimized implementation.** If we increase the token count from 1024 tokens to 10000 tokens, the original implementation will basically run forever without finishing.
- For most practical scenarios the performance impact is negligible. I tested with a book excerpt of 1800 tokens and both the original implementation and optimized implementation took ~0.00 seconds to run.
- Some practical scenarios - particularly those with long context and repetitive sequences - should significantly benefit from this PR. @Hunterius8 [noted](https://github.com/oobabooga/text-generation-webui/pull/5677#issuecomment-2028081501) that with 19k context the original DRY implementation was causing +50% slowdown to t/s. This optimized implementation should completely eliminate that slowdown.

## Root causes of the performance issues and how they were resolved

1. The original implementation loops over input ids in a torch tensor and accesses them like this: `input_ids_row[index].item()`. I discovered that torch adds a lot of overhead to this operation, making it extremely slow compared to the corresponding operation on a normal Python list. So we can trivially make the sampler a lot faster just by transforming the ids to a list. I isolated this change to commit 7c03e4af4d3eff so you can check that commit separately if you want. This change improves the performance on 1000 repetitive tokens to almost 0 seconds, and it improves the performance on 10000 repetitive tokens to 17 seconds. That's not good enough, so we need one more optimization:

2. The original implementation finds matching strings using a simple [O](https://en.wikipedia.org/wiki/Big_O_notation)(n^2) algorithm. For example, if the input size n=10000 repetitive tokens, then the string matching requires at least 10000 ^ 2 = 100 million steps (in the worst case). My optimized implementation finds matching strings using [Z algorithm](https://ivanyu.me/blog/2013/10/15/z-algorithm/), which is a faster [O](https://en.wikipedia.org/wiki/Big_O_notation)(n)  algorithm. Let's continue with the example of 10000 repetitive tokens to explain: O(n) means that instead of running 100 million steps, we only have to run 10000 steps - as the size of input increases, the number of steps increases linearly in relation to the size of the input. Since Z algorithm introduces some complexity in the code, I made an effort to write the algorithm as readable as possible and to document the steps. I also did extensive testing to verify correctness, since it's no longer obvious just by reading the code if it's correct or not.

## How to review this PR

I understand that reviewing first-time contributors' PRs is often annoying, so I made an effort to make the review process as easy as possible by attaching a testing helper file (temp1.py). You can run this file as a script independently of oobabooga (e.g. `python temp1.py`). It will run test cases where each test case is ran against both the original DRY implementation and the optimized DRY implementation. Running times are measured and results are compared. Error is thrown if results are different. If you look at the end of the temp1 file, you can see how you can add your own test cases there. You can, for example, [tokenize](https://belladoreai.github.io/llama3-tokenizer-js/example-demo/build/) a long prompt and copypaste the tokens into the temp1 file. At the end of the file you can also see a test generator, which generates random test data.

Note that I did almost all of my testing in the temp1 file, so it would be useful for someone to provide fresh eyes that the temp1 file corresponds to the actual changes in `sampler_hijack.py` and work there properly.

## Related observations

The following observations are not directly related to this PR, but they are observations I made about the DRY sampler while working on it:

- When running really large repetitive inputs, the sampler crashes due to OverflowError. My optimized implementation does not change the behavior of the sampler, so my implementation crashes with the same inputs as the original implementation crashes. This is trivial to fix, just let me know in the comments and I'll add it (I didn't add a fix yet because I wanted to be able to say "this PR improves performance without changing behavior").
- ~I think parameter `allowed_length` should be either renamed or its functionality should be changed, because the current behavior is "If the length of the matching sequence is less than allowed_length, no penalty is applied". For example, if allowed_length is 2, then it means the longest allowed length is 1 (not 2!). I expect this to confuse users if it's not changed.~
- I think sequence breakers parameter should be changed so that instead of strings, it takes token ids as input. Otherwise users will expect any string to act as a sequence breaker. The actual behavior is that strings are tokenized and some string might tokenize into multiple tokens, ~which if I understand correctly, will *each* act as a sequence breaker~ (edit: only last token of each string will be used). So if a really common token happens to end up in sequence breakers unbeknownst to the user, then it will hamper the behavior of the DRY sampler and the user might be dumbfounded as to why they are seeing so much repetition. By changing this to token ids we make it explicit to users what the inputs must be. 

## Checklist:

- [x] I have read the [Contributing guidelines](https://github.com/oobabooga/text-generation-webui/wiki/Contributing-guidelines).