@oobabooga 2024-05-23
> My optimized implementation finds matching strings using [Z algorithm](https://ivanyu.me/blog/2013/10/15/z-algorithm/), which is a faster [O](https://en.wikipedia.org/wiki/Big_O_notation)(n) algorithm.

It's really nice that you managed to reduce the time complexity of the algorithm, well done!

Some quick comments:

1) A simple sanity check is to paste some text in the Default or Notebook tabs and get the logits after sampling parameters in the dev branch and this PR branch, just to see if they are identical.
2) A fix to the OverflowError can be added as a new commit inside this PR. It would be good to run the sanity check above after this change.
3) `allowed_length` could perhaps be renamed to `minimum_penalized_length`. I'll tag the original author @p-e-w for his opinion.
4) Is it perhaps possible to use sequences of tokens as a sequence breaker rather than individual tokens? So that the problem you mentioned doesn't occur and the parameter becomes more literal.