@p-e-w 2024-05-24
Wow, that's awesome! A few initial comments:

:+1: for the list conversion, as that's a simple change with huge benefits that I overlooked. Thank you for finding and resolving this!

I'm less enthusiastic about the other change introducing the Z algorithm. Notably, I don't believe that performance on "adversarial" inputs such as 10k repeated tokens is a good justification for such a big algorithmic change. I spent literally hundreds of hours validating DRY in its current form, often stepping through the algorithm manually and quadruple-checking every single statement. I took a lot of care to produce an algorithm that is as easy to follow as possible, and rewrote everything multiple times in the process, because testing samplers is really, really hard and I wanted an implementation that can be easily ported to other loaders.

IMO, the proposed algorithm, while faster in pathological cases, is significantly more difficult to understand and reason about. It's great that you commented everything and even wrote tests to verify identical behavior to the current implementation, but unless the performance difference can be demonstrated to matter *with real-world inputs,* I don't think the benefits outweigh the additional complexity burden here. But of course, that's ultimately not my decision to make.

> When running really large repetitive inputs, the sampler crashes due to OverflowError.

Yes, I had anticipated that in theory and then forgotten about it. I believe it's a simple floating point overflow in the exponential penalty calculation. Note that this cannot occur unless the input already contains pathological repetitions, as DRY will prevent repetitive sequence buildups long before they can reach such lengths.

Clamping the exponent in the penalty formula should be sufficient to fix this.

> I think parameter `allowed_length` should be either renamed or its functionality should be changed, because the current behavior is "If the length of the matching sequence is less than allowed_length, no penalty is applied".

Nope, it's exactly what it says on the can. Here's the relevant code:

https://github.com/oobabooga/text-generation-webui/blob/852c9437690893bd953afacbb092eb929d139f32/modules/sampler_hijack.py#L285-L287

You're probably confused by the `>=` operator, which looks like the penalty would apply already to sequences that are equal in length to `allowed_length`. But note that penalties are always applied **to the token about to be sampled.** That is, we *already have* a repeated sequence of length `allowed_length`, so we penalize the next token in the probability distribution, because otherwise that token would make the repeated sequence *longer* than `allowed_length`, which is not allowed.

`allowed_length` is the maximum length that a repeated sequence can reach before penalties are applied *to tokens in that sequence.* As far as I'm concerned, this is what I would expect from the name. If `allowed_length` is 3, DRY will allow the input `one two three one two` to be continued without penalizing `three`, because `one two three` has length 3.

> I think sequence breakers parameter should be changed so that instead of strings, it takes token ids as input. Otherwise users will expect any string to act as a sequence breaker. The actual behavior is that strings are tokenized and some string might tokenize into multiple tokens, which if I understand correctly, will *each* act as a sequence breaker.

Not quite. Only the last token of each string is used, after prefixing the string with `a` in order to avoid problems with ambiguous tokenizations which exist particularly for punctuation.

Token IDs have the massive downside that they depend on the tokenizer, and there are quite a few tokenizers around now, with model makers continuing to experiment with vocabularies. I fear that expecting users to provide different token IDs every time they switch to another model would make this crucial parameter, and thus DRY itself, useless in practice, and above all else I want DRY to be practical to use.