@belladoreai 2024-05-24
Thank you both for such a quick response!

> 1. A simple sanity check is to paste some text in the Default or Notebook tabs and get the logits after sampling parameters in the dev branch and this PR branch, just to see if they are identical.

@oobabooga Can you please give more detailed instructions on how to run this sanity check?

It seems to me like the logits are not stable across generations. Even when I checkout a commit of the dev branch before my changes, and I generate with the exact same parameters and the exact same prompt, the logits after sampling are slightly different every time.

Here's an example of what I see when I try to print logits:

```
tensor([[[ 9.2031, 10.8047,  4.8984,  ..., -3.6758, -3.6758, -3.6777]]],
```

...and when I run the same thing again expecting to see the same logits, they are similar but not exactly the same:

```
tensor([[[ 9.2188, 10.7969,  4.9062,  ..., -3.6973, -3.6973, -3.6992]]]
```

I'm trying to print the logits inside file `exllamav2_hf.py` at the end of the `__call__` before this line is called:

```
return CausalLMOutputWithPast(logits=logits, past_key_values=seq if use_cache else None, loss=loss)
```

To me it seems like that would be the correct place to print the logits, but maybe you had some different place in mind?

Or maybe this issue is related to ExllamaV2 nondeterminism? I only have a 3060 so I'm unable to run any of the llama models without quantization.

Even if the logits were stable across generations, there would still be the problem of how to compare them. I can't eyeball through 128000 values, so I guess I would need to hash them and compare hashes or something? Maybe something like `hash(frozenset(logits[0][0].tolist()))`?

Alternatively, I also tried to check use the API endpoint to check logits: `http://127.0.0.1:5000/v1/internal/logits`. Something seems to be wrong with that endpoint. Making a call to it will hang forever and also causes the entire API to become unresponsive to all requests. Here's an example of one of the API calls I tried, which causes this issue:

```
{
    "prompt": "Who is best, Asuka or Rei? Answer:",
    "use_samplers": true,
    "top_k": 3,
    "max_tokens": 1,
    "max_new_tokens": 1
}
```