@belladoreai 2024-05-24
> 2. A fix to the OverflowError can be added as a new commit inside this PR. It would be good to run the sanity check above after this change.

I have now pushed a fix to the OverflowError as a separate commit.