@belladoreai 2024-05-24
> 3. allowed_length could perhaps be renamed to minimum_penalized_length

Sorry, this one is my bad. As @p-e-w explained, the current behavior is correct and no change to `allowed_length` is needed.